# distsql: Remove duplicate aggregator bucket maps