Rather than a `map[string]struct{}` map in the aggregator and a
`map[string]tree.AggregateFunc` per `aggregateFuncHolder` which have the
same set of keys, use a single `map[string][]tree.AggregateFunc`. This
reduces the number of map allocations required, which causes significant
overhead when the group columns are ordered and maps are frequently
allocated.

Here's a benchmark from `bench_test`:
```
name                                      old time/op    new time/op    delta
SortJoinAggregation/Cockroach-8             6.35ms ± 6%    5.70ms ± 5%  -10.16%  (p=0.000 n=10+10)
SortJoinAggregation/MultinodeCockroach-8    6.49ms ± 7%    5.44ms ± 2%  -16.16%  (p=0.000 n=9+10)

name                                      old alloc/op   new alloc/op   delta
SortJoinAggregation/Cockroach-8             2.61MB ± 1%    1.87MB ± 1%  -28.18%  (p=0.000 n=10+10)
SortJoinAggregation/MultinodeCockroach-8    2.68MB ± 1%    1.93MB ± 3%  -28.00%  (p=0.000 n=8+10)

name                                      old allocs/op  new allocs/op  delta
SortJoinAggregation/Cockroach-8              23.8k ± 1%     15.5k ± 2%  -34.77%  (p=0.000 n=10+10)
SortJoinAggregation/MultinodeCockroach-8     23.8k ± 1%     15.5k ± 3%  -34.86%  (p=0.000 n=8+10)
```

And here are some processor benchmarks of the identity function:
```
name                                 old time/op    new time/op    delta
Aggregation/IDENT-8                    58.7µs ± 3%    55.7µs ± 4%   -5.07%  (p=0.000 n=10+10)
OrderedAggregation/IDENT-8              230µs ± 1%     222µs ± 1%   -3.19%  (p=0.000 n=10+10)
PartiallyOrderedAggregation/IDENT-8     238µs ±10%     204µs ± 6%  -13.95%  (p=0.000 n=10+10)
UnorderedAggregation/IDENT-8            218µs ± 7%     174µs ± 2%  -19.82%  (p=0.000 n=10+10)

name                                 old speed      new speed      delta
Aggregation/IDENT-8                   136MB/s ± 3%   144MB/s ± 3%   +5.34%  (p=0.000 n=10+10)
OrderedAggregation/IDENT-8            105MB/s ± 1%   108MB/s ± 1%   +3.30%  (p=0.000 n=10+10)
PartiallyOrderedAggregation/IDENT-8   101MB/s ± 9%   118MB/s ± 6%  +16.14%  (p=0.000 n=10+10)
UnorderedAggregation/IDENT-8          110MB/s ± 6%   138MB/s ± 2%  +24.59%  (p=0.000 n=10+10)

name                                 old alloc/op   new alloc/op   delta
Aggregation/IDENT-8                    3.11kB ± 0%    2.98kB ± 0%   -4.11%  (p=0.000 n=10+10)
OrderedAggregation/IDENT-8             63.9kB ± 0%    52.5kB ± 0%  -17.87%  (p=0.000 n=10+10)
PartiallyOrderedAggregation/IDENT-8    25.1kB ± 0%    23.6kB ± 0%   -6.03%  (p=0.000 n=10+10)
UnorderedAggregation/IDENT-8           27.2kB ± 0%    25.2kB ± 0%   -7.21%  (p=0.000 n=10+9)

name                                 old allocs/op  new allocs/op  delta
Aggregation/IDENT-8                      17.0 ± 0%      16.0 ± 0%   -5.88%  (p=0.000 n=10+10)
OrderedAggregation/IDENT-8                719 ± 0%       519 ± 0%  -27.82%  (p=0.000 n=10+10)
PartiallyOrderedAggregation/IDENT-8       289 ± 0%       259 ± 0%  -10.38%  (p=0.000 n=10+10)
UnorderedAggregation/IDENT-8              241 ± 0%       230 ± 0%   -4.56%  (p=0.000 n=10+10)
```

Release note: None