@RaduBerinde 2018-04-18
Very nice! <img class="emoji" title=":lgtm_strong:" alt=":lgtm_strong:" align="absmiddle" src="https://reviewable.io/lgtm_strong.png" height="20" width="61"/> 

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24734#-:-LAPLNpWD-8VlyUZDYTX:b-il1ii9)*
<!-- Sent from Reviewable.io -->